requirements.txt
termcolor==2.3.0
aiohttp==3.8.5
ipdb==0.13.13
PyYAML==6.0
kubernetes==27.2.0
pandas==1.1.5
click==8.1.7